pyproject.toml
[tool.poetry]
name = "dbs"
version = "0.1.0"
description = ""
authors = ["dualdimension23 <bczypka@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cx-Oracle = "^8.3.0"
typenv = "^0.2.0"
requests = "^2.27.1"
PyYAML = "^6.0"
websockets = "^10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
